+++ app/assets/javascript/ahoy_captain/controllers/application_controller.js
@@ -2,6 +2,7 @@ import { Controller } from '@hotwired/stimulus';
 
 export default class extends Controller {
   connect() {
+    window.comboboxConnected = 0;
     if (new URLSearchParams(window.location.search).get('period') === 'realtime') {
       this.element.querySelectorAll('turbo-frame').forEach((frame) => {
         setInterval(() => {
@@ -10,4 +11,11 @@ export default class extends Controller {
       });
     }
   }
+
+  comboboxInit(event) {
+    if(event.detail.combobox.selectTarget.id === "property-name" || event.detail.combobox.selectTarget.id === "property-value") {
+      window.comboboxConnected += 1;
+    }
+  }
+
 }