+++ app/assets/javascript/ahoy_captain/controllers/tile_controller.js
@@ -1,0 +1,9 @@
+import { Controller } from '@hotwired/stimulus';
+
+export default class extends Controller {
+  static targets = ["title"]
+  setTitle(event) {
+    this.titleTarget.innerHTML = event.target.title || event.target.text
+  }
+
+}