+++ app/components/ahoy_captain/combobox_component.rb
@@ -1,0 +1,13 @@
+module AhoyCaptain
+  class ComboboxComponent < ViewComponent::Base
+    def initialize(name:, multiple: false, disabled: false, column:, url:, value: [], select_html: {})
+      @name = name
+      @multiple = multiple
+      @column = column
+      @url = url
+      @value = value
+      @select_html = select_html
+      @disabled = disabled
+    end
+  end
+end