+++ lib/ahoy_captain.rb
@@ -28,10 +28,10 @@ module AhoyCaptain
         pin "@hotwired/stimulus", to: "stimulus.min.js", preload: true
         pin "@hotwired/stimulus-loading", to: "stimulus-loading.js", preload: true
         pin "application", to: "ahoy_captain/application.js", preload: true
-        pin "slim-select", to: "https://ga.jspm.io/npm:slim-select@2.6.0/dist/slimselect.es.js", preload: true
         pin "chartkick", to: "chartkick.js"
         pin "Chart.bundle", to: "Chart.bundle.js"
         pin "chartjs-plugin-datalabels", to: "https://cdn.jsdelivr.net/npm/chartjs-plugin-datalabels@2", preload: true
+        pin "classnames", to: "https://cdnjs.cloudflare.com/ajax/libs/classnames/2.3.2/index.min.js", preload: true
         pin_all_from AhoyCaptain::Engine.root.join("app/assets/javascript/ahoy_captain/controllers"), under: "controllers", to: "ahoy_captain/controllers"
         pin_all_from AhoyCaptain::Engine.root.join("app/assets/javascript/ahoy_captain/helpers"), under: "helpers", to: "ahoy_captain/helpers"
       end